requirements.txt
openai>=1.0.0
pydantic>=2.0.0
matplotlib>=3.7.0
numpy>=1.24.0
python-dotenv>=1.0.0
rich>=13.0.0
pyyaml>=6.0.0
